Added more Network Packets, client features, and fixed a bug in Content Manager

+++ src/main/java/com/flansmod/client/model/GunAnimations.java
@@ -165,7 +165,7 @@ public class GunAnimations
         muzzleFlash = 2;
     }
 
-    public void doReload(int reloadTime, int pumpDelay, int pumpTime)
+    public void doReload(float reloadTime, int pumpDelay, int pumpTime)
     {
         reloading = true;
         lastReloadAnimationProgress = reloadAnimationProgress = 0F;
